chore: remove send module, add to client
